requirements.txt
awscli
requests
beautifulsoup4
pendulum
apache-airflow
tweepy
pandas
